requirements.txt
asgiref==3.2.10
Django==3.0.9
django-cors-headers==3.4.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
